Simplify chart generation
