+++ index.html
@@ -5,8 +5,9 @@
 <meta name="viewport" content="width=device-width,initial-scale=1">
 <style>
   html { font: 1rem/1.6 sans-serif; }
+  #chart { height: 480px; width: 960px; }
 </style>
 <h1>Is Puppeteer Firefox ready yet?</h1>
-<div id="output"></div>
 <div id="chart"></div>
+<script src="https://www.gstatic.com/charts/loader.js"></script>
 <script type="module" src="main.mjs"></script>
